Init API doc generation
