+++ babble-backend/index.js
@@ -1,13 +1,31 @@
 // // // // // // //
 //* SERVER INIT  *//
 // // // // // // //
+const express = require('express');
+const app = express();
+app.use(express.json());
 
 // on server start, populate db.js with persistent json
 const local = require('./data/db');
 local.loadData();
 
 // initialize routes
-require('./routes'); 
+const initRoutes = require('./routes');
+initRoutes(app, local);
+
+// API documentation generation
+const swaggerDoc = require('./swaggerDoc');
+swaggerDoc(app);
+
+// // // //  //
+//* DEPLOY  *//
+// // // //  //
+const port = 8080;
+
+app.listen(port);
+console.log(`listening on http://localhost:${port}/`);
+console.log(`API Docs at http://localhost:8080/docs`)
+console.log("Press Ctrl-C to quit");
 
 // // // //   //
 //* TESTING  *//
